requirements.txt
Django==4.0.2
djangorestframework==3.13.1
django-cors-headers==3.11.0
pymongo==3.12.1
djongo==1.3.6
